Make rust plugin lib global only
